Cleaned and tidied code. Added basic firestore service support.
