+++ pyrebase/types/keep_auth_session.py
@@ -1,0 +1,9 @@
+from requests import Session
+
+class KeepAuthSession(Session):
+    """
+    A session that doesn't drop Authentication on redirects between domains.
+    """
+
+    def rebuild_auth(self, prepared_request, response):
+        pass
