+++ pyrebase/types/pyre.py
@@ -1,0 +1,9 @@
+class Pyre:
+    def __init__(self, item):
+        self.item = item
+
+    def val(self):
+        return self.item[1]
+
+    def key(self):
+        return self.item[0]
